+++ packages/tracksu_storage/lib/tracksu_storage.dart
@@ -6,6 +6,7 @@ export 'src/flutter_secure_locale_store.dart';
 export 'src/flutter_secure_token_store.dart';
 export 'src/flutter_secure_oauth_transaction_store.dart';
 export 'src/locale_store.dart';
+export 'src/theme_store.dart';
 export 'src/oauth_transaction_store.dart';
 export 'src/pending_oauth_transaction.dart';
 export 'src/stored_auth_tokens.dart';